+++ tests/template_hdf5.c
@@ -1,0 +1,5 @@
+#define TEST_BACKEND_HDF5
+#define TREXIO_FILE_PREFIX "REPLACE"
+#include "test_macros.h"
+#include "REPLACE.c"
+
