+++ tests/template_text.c
@@ -1,0 +1,5 @@
+#define TEST_BACKEND_TEXT
+#define TREXIO_FILE_PREFIX "REPLACE"
+#include "test_macros.h"
+#include "REPLACE.c"
+
